+++ node/src/components/chain_synchronizer/operations.rs
@@ -694,7 +694,21 @@ async fn sync_trie_store(state_root_hash: Digest, ctx: &ChainSyncContext<'_>) ->
 
     let queue = Arc::new(WorkQueue::default());
     queue.push_job(state_root_hash);
-    let mut workers: FuturesUnordered<_> = (0..ctx.config.max_parallel_trie_fetches())
+    let peer_count = cmp::max(
+        1,
+        ctx.effect_builder
+            .get_fully_connected_peers()
+            .await
+            .len()
+            .saturating_sub(
+                ctx.bad_peer_list
+                    .read()
+                    .expect("bad peer list lock poisoned")
+                    .len(),
+            ),
+    );
+    let parallel_count = ctx.config.max_parallel_trie_fetches_per_peer() * peer_count;
+    let mut workers: FuturesUnordered<_> = (0..parallel_count)
         .map(|worker_id| sync_trie_store_worker(worker_id, abort.clone(), queue.clone(), ctx))
         .collect();
     while let Some(result) = workers.next().await {
@@ -1377,9 +1391,23 @@ async fn fetch_and_store_deploys(
 
     let hashes: Vec<_> = hashes.cloned().collect();
     let mut deploys: Vec<Deploy> = Vec::with_capacity(hashes.len());
+    let peer_count = cmp::max(
+        1,
+        ctx.effect_builder
+            .get_fully_connected_peers()
+            .await
+            .len()
+            .saturating_sub(
+                ctx.bad_peer_list
+                    .read()
+                    .expect("bad peer list lock poisoned")
+                    .len(),
+            ),
+    );
+    let parallel_count = ctx.config.max_parallel_deploy_fetches_per_peer() * peer_count;
     let mut stream = futures::stream::iter(hashes)
         .map(|hash| fetch_and_store_deploy(hash, ctx))
-        .buffer_unordered(ctx.config.max_parallel_deploy_fetches());
+        .buffer_unordered(parallel_count);
     while let Some(result) = stream.next().await {
         let deploy = result?;
         trace!("fetched {:?}", deploy);