+++ node/src/types/node_config.rs
@@ -3,10 +3,12 @@ use serde::{Deserialize, Serialize};
 
 use crate::types::{BlockHash, TimeDiff};
 
-/// Maximum number of deploys to fetch in parallel, by default.
-const DEFAULT_MAX_PARALLEL_DEPLOY_FETCHES: u32 = 20;
-/// Maximum number of tries to fetch in parallel, by default.
-const DEFAULT_MAX_PARALLEL_TRIE_FETCHES: u32 = 20;
+/// Maximum number of fetch-deploy tasks per peer to run in parallel during chain synchronization,
+/// by default.
+const DEFAULT_MAX_PARALLEL_DEPLOY_FETCHES_PER_PEER: u32 = 10;
+/// Maximum number of fetch-trie tasks per peer to run in parallel during chain synchronization, by
+/// default.
+const DEFAULT_MAX_PARALLEL_TRIE_FETCHES_PER_PEER: u32 = 10;
 const DEFAULT_PEER_REDEMPTION_INTERVAL: u32 = 10_000;
 const DEFAULT_RETRY_INTERVAL: &str = "100ms";
 
@@ -18,11 +20,13 @@ pub struct NodeConfig {
     /// Hash used as a trust anchor when joining, if any.
     pub trusted_hash: Option<BlockHash>,
 
-    /// Maximum number of deploys to fetch in parallel.
-    pub max_parallel_deploy_fetches: u32,
+    /// Maximum number of fetch-deploy tasks per peer to run in parallel during chain
+    /// synchronization.
+    pub max_parallel_deploy_fetches_per_peer: u32,
 
-    /// Maximum number of trie nodes to fetch in parallel.
-    pub max_parallel_trie_fetches: u32,
+    /// Maximum number of fetch-trie tasks per peer to run in parallel during chain
+    /// synchronization.
+    pub max_parallel_trie_fetches_per_peer: u32,
 
     /// The duration for which to pause between retry attempts while synchronising during joining.
     pub retry_interval: TimeDiff,
@@ -39,8 +43,8 @@ impl Default for NodeConfig {
     fn default() -> NodeConfig {
         NodeConfig {
             trusted_hash: None,
-            max_parallel_deploy_fetches: DEFAULT_MAX_PARALLEL_DEPLOY_FETCHES,
-            max_parallel_trie_fetches: DEFAULT_MAX_PARALLEL_TRIE_FETCHES,
+            max_parallel_deploy_fetches_per_peer: DEFAULT_MAX_PARALLEL_DEPLOY_FETCHES_PER_PEER,
+            max_parallel_trie_fetches_per_peer: DEFAULT_MAX_PARALLEL_TRIE_FETCHES_PER_PEER,
             retry_interval: DEFAULT_RETRY_INTERVAL.parse().unwrap(),
             sync_peer_redemption_interval: DEFAULT_PEER_REDEMPTION_INTERVAL,
             sync_to_genesis: false,